Refactoring: WriteOnReadTracing - move reusable parts out of browser view:
This is in preparation for using them to (optionally) include tracebacks
of the *actual* writes that caused the problem in the ReadOnlyError
error page, not just the TB of the location where the error was raised.
